requirements.txt
# Python: 3.10–3.13 (see pyproject.toml requires-python)
chainlit>=2.0.0,<3
langchain>=0.3.14,<0.4
langchain-community>=0.3.14,<0.4
# Keep core on 0.3.x with langchain 0.3 (avoids accidental langchain-core 1.x from other tools)
langchain-core>=0.3.14,<0.4
langchain-text-splitters>=0.3.0,<0.4
groq>=0.12.0
python-dotenv>=1.0.0
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-pptx>=0.6.21
sentence-transformers>=3.3.0,<4
faiss-cpu>=1.8.0
requests>=2.28.0
pytest>=8.0.0
